pi-voice-assistant/backend/requirements.txt
fastapi
uvicorn
ollama
faster-whisper
piper-tts
python-multipart
pyaudio
websockets
vosk
stt
st7789
spidev
Pillow
